requirements.txt
asgiref==3.10.0
attrs==25.4.0
autobahn==25.10.2
Automat==25.4.16
bleach==6.3.0
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.95.0
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
iniconfig==2.3.0
msgpack==1.1.2
packaging==25.0
pluggy==1.6.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pytest==9.0.1
pytest-django==4.11.1
python-decouple==3.8
redis==7.0.1
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
webencodings==0.5.1
zope.interface==8.1
